refactor: move repository e controller sem paginacao para modulo da table
