+++ src/features/acf/backend/hypertension/index.ts
@@ -3,11 +3,8 @@ export {
     pageCoaps as getPageCoaps,
     rowCountCoaps as getRowCountCoaps,
     rowCountCoeq as getRowCountCoeq,
+    allDataCoaps as getAllDataCoaps,
+    allDataCoeq as getAllDataCoeq,
 } from "./modules/table";
 
 export { filterOptionsCoeq, filterOptionsCoaps } from "./modules/filters";
-
-export {
-    dataPrintCoaps as getDataPrintCoaps,
-    dataPrintCoeq as getDataPrintCoeq,
-} from "./modules/print";
