--- src/features/acf/backend/hypertension/modules/print/index.ts
@@ -1,1 +1,0 @@
-export { dataPrintCoeq, dataPrintCoaps } from "./controller";
