--- src/features/acf/backend/hypertension/modules/print/model.ts
@@ -1,3 +1,0 @@
-import type * as model from "@/features/acf/shared/hypertension/model";
-
-export type PageItem = Omit<model.HypertensionAcfItem, "patientAgeRange">;
