+++ src/features/acf/backend/hypertension/modules/table/index.ts
@@ -1,1 +1,8 @@
-export { pageCoeq, pageCoaps, rowCountCoaps, rowCountCoeq } from "./controller";
+export {
+    pageCoeq,
+    pageCoaps,
+    rowCountCoaps,
+    rowCountCoeq,
+    allDataCoaps,
+    allDataCoeq,
+} from "./controller";
